move updateMode and ModeState into utils
this avoids circular imports. utils may not be the perfect place, but much better than the circular imports
