+++ web/src/js/ducks/modes/utils.ts
@@ -1,4 +1,39 @@
-import { ModeState } from "../modes";
+import {getMode as getRegularModeConfig} from "./regular";
+import {getMode as getLocalModeConfig} from "./local";
+import {fetchApi} from "../../utils";
+
+export interface ModeState {
+    active: boolean;
+    name: string;
+    listen_port?: number;
+    listen_host?: string;
+    error?: string;
+}
+
+export const updateMode = () => {
+    return async (_, getState) => {
+        try {
+            const modes = getState().modes;
+
+            const activeModes: string[] = [
+                ...getRegularModeConfig(modes),
+                ...getLocalModeConfig(modes),
+                //add new modes here
+            ];
+            const response = await fetchApi.put("/options", {
+                mode: activeModes,
+            });
+            if (response.status === 200) {
+                return { success: true };
+            } else {
+                const errorText = await response.text();
+                return { success: false, error: errorText };
+            }
+        } catch (error) {
+            //TODO: handle error
+        }
+    };
+};
 
 export const addListenAddr = (mode: ModeState) => {
     let stringMode = mode.name;
